Import all xtream data

+++ src/main/kotlin/io/github/firstred/iptvproxy/db/tables/epg/EpgProgrammeAudioTable.kt
@@ -5,7 +5,7 @@ import org.jetbrains.exposed.sql.ReferenceOption
 import org.jetbrains.exposed.sql.Table
 
 object EpgProgrammeAudioTable : Table("epg_programme_audio") {
-    val programmeId: Column<Int> = integer("programme_id")
+    val programmeId: Column<Long> = long("programme_id")
     val server: Column<String> = varchar("server", 511)
     val type: Column<String> = varchar("type", 511)
     val value: Column<String> = varchar("value", 511)
